# ENH: Implement sindex query in clip