--- a/geopandas/tools/clip.py
+++ b/geopandas/tools/clip.py
@@ -36,12 +36,7 @@ def _clip_points(gdf, poly):
         The returned GeoDataFrame is a subset of gdf that intersects
         with poly.
     """
-    spatial_index = gdf.sindex
-    bbox = poly.bounds
-    sidx = list(spatial_index.intersection(bbox))
-    gdf_sub = gdf.iloc[sidx]
-
-    return gdf_sub[gdf_sub.geometry.intersects(poly)]
+    return gdf.iloc[gdf.sindex.query(poly, predicate="intersects")]
 
 
 def _clip_line_poly(gdf, poly):
@@ -65,14 +60,7 @@ def _clip_line_poly(gdf, poly):
         The returned GeoDataFrame is a clipped subset of gdf
         that intersects with poly.
     """
-    spatial_index = gdf.sindex
-
-    # Create a box for the initial intersection
-    bbox = poly.bounds
-    # Get a list of id's for each object that overlaps the bounding box and
-    # subset the data to just those lines
-    sidx = list(spatial_index.intersection(bbox))
-    gdf_sub = gdf.iloc[sidx]
+    gdf_sub = gdf.iloc[gdf.sindex.query(poly, predicate="intersects")]
 
     # Clip the data with the polygon
     if isinstance(gdf_sub, GeoDataFrame):
@@ -80,11 +68,11 @@ def _clip_line_poly(gdf, poly):
         clipped["geometry"] = gdf_sub.intersection(poly)
 
         # Return the clipped layer with no null geometry values or empty geometries
-        return clipped[~clipped.geometry.is_empty & clipped.geometry.notnull()]
+        return clipped
     else:
         # GeoSeries
         clipped = gdf_sub.intersection(poly)
-        return clipped[~clipped.is_empty & clipped.notnull()]
+        return clipped
 
 
 def clip(gdf, mask, keep_geom_type=False):