Quick PR to implement the new spatial index capabilities in clip.