@jorisvandenbossche 2020-05-15
Cool! That nicely shows the benefits of the improved query interface ;)