@adriangb 2020-05-15
I have not run any benchmarks. However, based on the benchmarks I have run on just `query_bulk` and `sindex.intersection` I think this will be faster. Let me do some quick testing to confirm.