@adriangb 2020-05-15
Benchmarks added. Results:
| before     | after     | ratio | pygeos installed?    |   |
|------------|-----------|-------|----------------------|---|
| 1.37±0.09s | 927±40ms  | ~0.68 | clip.Bench.time_clip | Y |
| 3.03±0.1s  | 2.98±0.2s | 0.98  | clip.Bench.time_clip | N |