@adriangb 2020-05-15
Might want to restart this job @martinfleis : https://travis-ci.org/github/geopandas/geopandas/jobs/687544175